aggtiff/requirements.txt
requests
python-dateutil
simplejson
